hook cancelled. Validate/state separated. Now working as expected
